# lib/api-zod/src/generated/types/lessonProgressList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonProgress } from './lessonProgress';

export interface LessonProgressList {
  items: LessonProgress[];
}