# lib/api-zod/src/generated/types/preferenceList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Preference } from './preference';

export interface PreferenceList {
  items: Preference[];
}