# lib/api-zod/src/generated/types/quizResultList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizResult } from './quizResult';

export interface QuizResultList {
  items: QuizResult[];
}