# lib/api-zod/src/generated/types/savedItemList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SavedItem } from './savedItem';

export interface SavedItemList {
  items: SavedItem[];
}